requirements.txt
langchain
langchain-community
langchain-ollama
ollama
streamlit
faiss-cpu
PyPDF2
transformers